Add common collector
DRY up collector codebase with common collector.
Add New Relic to collectors
